+++ app/CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Light and dark theme.
 - Option to bind multiple keys/buttons to the same input.
 - New color palettes: SONY CXA2025AS US, Unsaturated V6.
+- **Server:** Configuration through file or environment variables.
+- **Server:** HTTPS and (experimental) HTTP2 support.
 
 ### Changed